NOTE: BIG COMMIT - COMPILER WARNINGS
Removing alot of compiler warnings. This has been on my to-do list for quite some time. Wanted everything in one big commit - so this is it...

Some of the 3rd party header libraries do have warnings - these have been explicitly turned off in the CMakeLists file.

+++ src/Core/Controllers/BaseController.cpp
@@ -42,7 +42,7 @@ bool BaseController::DefaultEditSpecial(Cursor &cursor, Line::Ref line, const Ke
                 wasHandled = true;
                 break;
             case Keyboard::kKeyCode_DeleteForward :
-                if (cursor.position.x < line->Length()) {
+                if (cursor.position.x < (int)line->Length()) {
                     line->Delete(cursor.position.x);
                     wasHandled = true;
                 }
